work on bev_detect_model(1)
