+++ BEV/yolo/dataset.yaml
@@ -7,5 +7,5 @@ nc: 2
 
 # class names
 names:
-  0: other
-  1: car
+  0: other car
+  1: emergency car
