+++ BEV/yolo/runs/obb/train3/results.csv
@@ -1,0 +1,24 @@
+epoch,time,train/box_loss,train/cls_loss,train/dfl_loss,metrics/precision(B),metrics/recall(B),metrics/mAP50(B),metrics/mAP50-95(B),val/box_loss,val/cls_loss,val/dfl_loss,lr/pg0,lr/pg1,lr/pg2
+1,746.685,1.94034,2.17095,2.37776,0.88275,0.59482,0.73096,0.42401,1.5534,1.64516,2.10467,0.00332889,0.00332889,0.00332889
+2,1491.05,1.27047,0.96645,2.01999,0.88545,0.76285,0.85269,0.56813,1.31184,1.34526,1.93035,0.00659627,0.00659627,0.00659627
+3,2234.66,1.13889,0.83338,1.87506,0.88639,0.76495,0.86921,0.60419,1.23838,1.21005,1.94259,0.00979764,0.00979764,0.00979764
+4,2976.16,1.05267,0.76217,1.81971,0.89595,0.7987,0.89556,0.63316,1.22036,1.2737,1.92871,0.009703,0.009703,0.009703
+5,3717.34,0.97637,0.70384,1.74775,0.91244,0.81874,0.91239,0.59321,1.19259,1.6528,2.00277,0.009604,0.009604,0.009604
+6,4460.86,0.91656,0.65517,1.71515,0.85697,0.84047,0.89755,0.57983,1.19996,1.18676,1.93238,0.009505,0.009505,0.009505
+7,5206.26,0.88388,0.6327,1.6721,0.91853,0.83403,0.92471,0.60765,1.17032,0.96119,2.03471,0.009406,0.009406,0.009406
+8,5953.87,0.86776,0.61882,1.65363,0.94233,0.85276,0.93149,0.62697,1.09986,0.80806,1.91268,0.009307,0.009307,0.009307
+9,6694.84,0.82745,0.59316,1.63704,0.90565,0.793,0.88691,0.68759,1.01931,1.04642,1.85457,0.009208,0.009208,0.009208
+10,7435.42,0.81036,0.58524,1.60688,0.88187,0.83759,0.88707,0.64182,1.09292,0.84217,2.00716,0.009109,0.009109,0.009109
+11,8175.93,0.78126,0.5587,1.60561,0.92765,0.84846,0.9018,0.63193,1.06164,0.89844,2.06987,0.00901,0.00901,0.00901
+12,8917.43,0.76952,0.55251,1.58932,0.90664,0.81366,0.91232,0.66381,1.04393,1.0013,2.00232,0.008911,0.008911,0.008911
+13,9658.57,0.75965,0.54709,1.57398,0.91334,0.83777,0.8886,0.65906,1.04593,0.73221,1.96996,0.008812,0.008812,0.008812
+14,10399.5,0.74411,0.53351,1.56684,0.89669,0.81644,0.86059,0.65113,0.9824,0.76417,1.95556,0.008713,0.008713,0.008713
+15,11142.6,0.7283,0.51924,1.567,0.9457,0.83935,0.89823,0.66673,1.02913,0.77718,1.89965,0.008614,0.008614,0.008614
+16,11902,0.71465,0.50929,1.55844,0.90161,0.86333,0.91149,0.67172,1.0476,0.8097,2.02959,0.008515,0.008515,0.008515
+17,12655.5,0.70449,0.5059,1.54937,0.98019,0.84885,0.91801,0.68895,1.00864,0.67629,1.98395,0.008416,0.008416,0.008416
+18,13422.7,0.69647,0.49548,1.5385,0.9316,0.83753,0.9099,0.68228,1.02235,0.70605,2.0119,0.008317,0.008317,0.008317
+19,14167.8,0.68857,0.49298,1.53231,0.93506,0.85475,0.91915,0.69989,0.97698,0.72848,2.01053,0.008218,0.008218,0.008218
+20,14912.5,0.67948,0.48353,1.52875,0.97343,0.862,0.92761,0.6919,1.00038,0.65186,1.9477,0.008119,0.008119,0.008119
+21,15664.8,0.67508,0.48706,1.50365,0.90785,0.86544,0.92536,0.68379,0.96657,0.65992,1.9398,0.00802,0.00802,0.00802
+22,16425.8,0.66226,0.47443,1.51579,0.96616,0.83262,0.90994,0.68664,0.97786,0.68995,1.95251,0.007921,0.007921,0.007921
+23,17220.4,0.65882,0.4692,1.51243,0.94732,0.82594,0.89805,0.70189,0.97902,0.67018,1.92334,0.007822,0.007822,0.007822
